service/frontend/streamlit-data-app/requirements.txt
streamlit
requests
pandas
numpy
matplotlib
stop-words
nltk
python-dotenv
openpyxl
xlrd
pyxlsb
odfpy